refact: removes sign in screen rules and liability
