+++ src/components/CustomTextInput/styles.ts
@@ -1,3 +1,4 @@
+import { Feather } from '@expo/vector-icons';
 import styled from 'styled-components/native';
 
 export const Container = styled.View`
@@ -35,10 +36,10 @@ export const Label = styled.Text<{ isFocused: boolean }>`
   transition: top 0.2s ease, font-size 0.2s ease;
 `;
 
-export const LeftIcon = styled`
+export const LeftIcon = styled(Feather)`
   margin-right: 10px;
 `;
 
-export const ClearIcon = styled`
+export const ClearIcon = styled(Feather)`
   margin-left: 10px;
 `;
