+++ src/pages/SignInPage/index.tsx
@@ -1,43 +1,16 @@
 import CustomTextInput from '@components/CustomTextInput'
 import { ImageLogo } from '@components/Logo'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useAuth } from '@hooks/useAuth'
-import { useNetworkStatus } from '@hooks/useNetworkStatus'
 import { useThemeSwitcher } from '@hooks/useThemeSwitcher'
-import React, { useEffect, useState } from 'react'
-import { Controller, useForm } from 'react-hook-form'
+import React, { useEffect } from 'react'
+import { Controller } from 'react-hook-form'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
-import { type AuthFormData, authSchema } from 'src/libs/zod/user'
 
 import * as S from './styles'
+import { useSignInViewModel } from './viewModel'
 
 export const SignInPage = () => {
   const { theme } = useThemeSwitcher()
-  const { login } = useAuth();
-  const isConnected = useNetworkStatus();
-
-  const [error, setError] = useState<string | null>(null)
-
-  const form = useForm<AuthFormData>({
-    resolver: zodResolver(authSchema),
-  })
-
-  const handleSubmit = (data: AuthFormData) => {
-    try {
-      if (data.username === 'user' && data.password === '123') {
-        if (isConnected) {
-          login();
-        } else {
-          setError('Sem conexão com a internet');
-        }
-      }
-      else {
-        setError('Credenciais inválidas')
-      }
-    } catch (error: any) {
-      setError(error.message)
-    }
-  }
+  const { error, form, setError, handleSubmit } = useSignInViewModel();
 
   useEffect(() => {
     if (error !== null) {
